requirements/base.txt
pytz==2024.1  # https://github.com/stub42/pytz
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==10.3.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.6.0  # https://github.com/evansd/whitenoise
redis==5.0.3  # https://github.com/redis/redis-py
hiredis==2.3.2  # https://github.com/redis/hiredis-py
uvicorn[standard]==0.20.0  # https://github.com/encode/uvicorn
rcssmin==1.1.1  # https://github.com/ndparker/rcssmin

# Django
# ------------------------------------------------------------------------------
django==4.2.8  # LTS https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.5.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.61.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-compressor==4.3.1  # https://github.com/django-compressor/django-compressor
django-redis==5.2.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.15.1  # https://github.com/encode/django-rest-framework
django-cors-headers==4.3.1 # https://github.com/adamchainz/django-cors-headers
coreapi==2.3.3 # Included explicitly since it's defined optional by djangorestframework
# DRF-spectacular for api documentation
drf-spectacular==0.27.2  # https://github.com/tfranzel/drf-spectacular
django-q2==1.7.2

# Utils
# ------------------------------------------------------------------------------
pytz
colander-data-converter

# Vendors
# ------------------------------------------------------------------------------
OTXv2
vt-py
shodan==1.31.0
scarlet-shark-client>=1.0.6
